% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{after_submit}
\alias{after_submit}
\title{Provide a message with context about a path variable, after submitting}
\usage{
after_submit(repo, path, browse_pr = FALSE)
}
\arguments{
\item{repo}{Local path to repo}

\item{path}{Path of knowledge post}

\item{browse_pr}{Whether to browse to a pull request}
}
\description{
Provide a message with context about a path variable, after submitting
}
