% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_remote_link}
\alias{get_remote_link}
\title{Retrieve the remote link of a GitHub repository}
\usage{
get_remote_link(path)
}
\arguments{
\item{path}{Path to a local git repository}
}
\description{
This is useful for linking to a remote pull request
}
