% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{kr_add}
\alias{kr_add}
\title{Add a post to a knowledge repository}
\usage{
kr_add(filename, path = NULL, update = FALSE, branch = NULL,
  squash = FALSE, submit = FALSE, message = paste0("Adding ", filename),
  src = NULL, ...)
}
\arguments{
\item{filename}{Filename to add}

\item{path}{The path of the destination post to be added in the
knowledge repo. Required if the knowledge post does
not specify "path" in its headers.}

\item{update}{Whether this should update an existing post of the
same name.}

\item{branch}{The branch to use for this addition, if not the
default (which is the path of the knowledge post).}

\item{squash}{Automatically suppress all previous commits, and
replace it with this version.}

\item{submit}{Submit newly added post}

\item{message}{Commit message}

\item{src}{Specify additional source files to add to
\code{<knowledge_post>/orig_src.}}

\item{...}{Global arguments to \code{knowledge_repo} commands, such as
\code{repo} or \code{noupdate} (see Details)}
}
\description{
Add a post to a knowledge repository
}
\details{
Global arguments that can be passed to all knowledge_repo
commands include:

\itemize{
  \item{repo}{Path to Knowledge Repository to add to. Can also set
  \code{KNOWLEDGE_REPO} variable in \code{.Renviron}.}
  \item{dev}{Whether to skip passing control to version of code
  checked out in knowledge repository.}
  \item{debug}{Whether to enable debug mode.}
  \item{noupdate}{Whether script should update the repository before
  performing actions.}
  \item{version}{Show version and exit.}
  \item{help}{Show help and exit}
}
}
\examples{

kr_add("test.Rmd")
kr_add("test.Rmd", message = "Committing a new post")
kr_add("test.Rmd", submit = TRUE)

}
