% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{kr_command}
\alias{kr_command}
\title{Run a keybase command on the command line}
\usage{
kr_command(..., .verbose = TRUE)
}
\arguments{
\item{...}{Arguments to construct a command, either unnamed (positional)
or named}

\item{.verbose}{Whether to display the output as a message}
}
\description{
Run a keybase command with the given arguments. Positional
arguments are provided in order, while named arguments are prefixed
with \code{--name}. Arguments passed as TRUE are empty (e.g.
\code{help = TRUE} becomes \code{--help}), while FALSE and NULL are
dropped.
}
\examples{

kr_command("create", "Rmd", "test.Rmd")

}
