% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{kr_deploy}
\alias{kr_deploy}
\title{Deploy a local server}
\usage{
kr_deploy(port = NULL, dburi = NULL, workers = NULL, timeout = NULL,
  config = NULL, engine = NULL, ...)
}
\arguments{
\item{port}{Specify the port on which to run the web server.}

\item{dburi}{The SQLAlchemy database uri.}

\item{workers}{Number of gunicorn worker threads to spin up.}

\item{timeout}{Specify the timeout (seconds) for the gunicorn web
server.}

\item{config}{The config file from which to read server configuration.}

\item{engine}{Which server engine to use when deploying; choose
from: "flask", "gunicorn" (default) or "uwsgi".}

\item{...}{Global arguments to \code{knowledge_repo} commands, such as
\code{repo} or \code{noupdate} (see Details)}
}
\description{
Deploy a local server
}
\details{
Global arguments that can be passed to all knowledge_repo
commands include:

\itemize{
  \item{repo}{Path to Knowledge Repository to add to. Can also set
  \code{KNOWLEDGE_REPO} variable in \code{.Renviron}.}
  \item{dev}{Whether to skip passing control to version of code
  checked out in knowledge repository.}
  \item{debug}{Whether to enable debug mode.}
  \item{noupdate}{Whether script should update the repository before
  performing actions.}
  \item{version}{Show version and exit.}
  \item{help}{Show help and exit}
}
}
\examples{

kr_deploy()

}
