% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{kr_init}
\alias{kr_init}
\title{Initialize a new knowledge repository}
\usage{
kr_init(repo, tooling_embed = FALSE, tooling_repo = NULL,
  tooling_branch = NULL, ...)
}
\arguments{
\item{repo}{Folder of the repository to create}

\item{tooling_embed}{Embed a reference version knowledge_repo tooling in
the repository.}

\item{tooling_repo}{The repository to use (if not the default).}

\item{tooling_branch}{The branch to use when embedding the tools as a
submodule (default is "master").}

\item{...}{Global arguments to \code{knowledge_repo} commands, such as
\code{repo} or \code{noupdate} (see Details)}
}
\description{
Initialize a new knowledge repository
}
\details{
Global arguments that can be passed to all knowledge_repo
commands include:

\itemize{
  \item{repo}{Path to Knowledge Repository to add to. Can also set
  \code{KNOWLEDGE_REPO} variable in \code{.Renviron}.}
  \item{dev}{Whether to skip passing control to version of code
  checked out in knowledge repository.}
  \item{debug}{Whether to enable debug mode.}
  \item{noupdate}{Whether script should update the repository before
  performing actions.}
  \item{version}{Show version and exit.}
  \item{help}{Show help and exit}
}
}
\examples{

kr_init("example_repository")

}
