% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{kr_preview}
\alias{kr_preview}
\title{Preview a post locally}
\usage{
kr_preview(path, port = NULL, dburi = NULL, config = NULL, ...)
}
\arguments{
\item{path}{The path of the knowledge post to preview.}

\item{port}{Specify the port on which to run the web server.}

\item{dburi}{The SQLAlchemy database uri.}

\item{config}{Configuration file}

\item{...}{Global arguments to \code{knowledge_repo} commands, such as
\code{repo} or \code{noupdate} (see Details)}
}
\description{
Preview a post locally
}
\details{
Global arguments that can be passed to all knowledge_repo
commands include:

\itemize{
  \item{repo}{Path to Knowledge Repository to add to. Can also set
  \code{KNOWLEDGE_REPO} variable in \code{.Renviron}.}
  \item{dev}{Whether to skip passing control to version of code
  checked out in knowledge repository.}
  \item{debug}{Whether to enable debug mode.}
  \item{noupdate}{Whether script should update the repository before
  performing actions.}
  \item{version}{Show version and exit.}
  \item{help}{Show help and exit}
}
}
