% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{kr_status}
\alias{kr_status}
\title{Show status of knowledge repo}
\usage{
kr_status(...)
}
\arguments{
\item{...}{Global arguments to \code{knowledge_repo} commands, such as
\code{repo} or \code{noupdate}

kr_status()}
}
\description{
Show status of knowledge repo
}
