% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{kr_command}
\alias{kr_command}
\title{Run a keybase command on the command line}
\usage{
kr_command(..., .verbose = TRUE)
}
\arguments{
\item{...}{Arguments to construct a command, either unnamed (positional)
or named}

\item{.verbose}{Whether to display the output as a message}

\item{...}{Global arguments to \code{knowledge_repo} commands, such as
\code{repo} or \code{noupdate} (see Details)}
}
\description{
Run a keybase command with the given arguments. Positional
arguments are provided in order, while named arguments are prefixed
with \code{--name}. To make passing arguments easier, commit messages
}
\details{
Global arguments that can be passed to all knowledge_repo
commands include:

\itemize{
  \item{repo}{Path to Knowledge Repository to add to. Can also set
  \code{KNOWLEDGE_REPO} variable in \code{.Renviron}.}
  \item{dev}{Whether to skip passing control to version of code
  checked out in knowledge repository.}
  \item{debug}{Whether to enable debug mode.}
  \item{noupdate}{Whether script should update the repository before
  performing actions.}
  \item{version}{Show version and exit.}
  \item{help}{Show help and exit}
}
}
\examples{

kr_command("create", "Rmd", "test.Rmd")
kr_command("add", help = TRUE)

}
