% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{kr_submit}
\alias{kr_submit}
\title{Submit a post to the knowledge base for review}
\usage{
kr_submit(path, ...)
}
\arguments{
\item{path}{The path of the knowledge post to submit for review.}

\item{...}{Global arguments to \code{knowledge_repo} commands, such as
\code{repo} or \code{noupdate} (see Details)}
}
\description{
Submit a post to the knowledge base for review. This would be done
after adding it (assuming submit was FALSE in that command).
}
\details{
Global arguments that can be passed to all knowledge_repo
commands include:

\itemize{
  \item{repo}{Path to Knowledge Repository to add to. Can also set
  \code{KNOWLEDGE_REPO} variable in \code{.Renviron}.}
  \item{dev}{Whether to skip passing control to version of code
  checked out in knowledge repository.}
  \item{debug}{Whether to enable debug mode.}
  \item{noupdate}{Whether script should update the repository before
  performing actions.}
  \item{version}{Show version and exit.}
  \item{help}{Show help and exit}
}
}
\examples{

kr_create("test.Rmd")
kr_add("test.Rmd", path = "examples/test")
kr_submit("examples/test")

}
